Add files via upload
Newton numerical method. It is to notice that for the exponential function we get something 5E-17. This is "very small" and coherent with the theoretical result. However, this stabilized value (in the limits of the performance of the machine) is still non-zero value. Theoretically, the convergence is of order at least 1, but it seems that in some degree it becomes too slow. what explains this situation ?